Collapses ForSale and NotForSale artwork grids.
