+++ src/lib/Components/ArtworkGrids/RelayConnections/__tests__/ArtistArtworksGrid-tests.tsx
@@ -3,8 +3,8 @@ import "react-native"
 import React from "react"
 import { renderWithLayout } from "../../../../tests/renderWithLayout"
 
+import ArtistArtworksGridContainer from "../../../Artist/Artworks"
 import artwork from "../../__tests__/__fixtures__/artwork"
-import { ArtistForSaleArtworksGridContainer } from "../ArtistForSaleArtworksGrid"
 
 // FIXME: This needs a Relay based test
 xit("renders properly", () => {
@@ -14,6 +14,6 @@ xit("renders properly", () => {
     },
   }
 
-  const grid = renderWithLayout(<ArtistForSaleArtworksGridContainer artist={artist as any} />, { width: 768 })
+  const grid = renderWithLayout(<ArtistArtworksGridContainer artist={artist as any} />, { width: 768 })
   expect(grid).toMatchSnapshot()
 })
